+++ tools/bbcorenpm/index.ts
@@ -64,7 +64,7 @@ async function callRepoApi(path: string) {
         proxy: process.env.http_proxy || process.env.https_proxy,
         headers: {
             accept: "application/vnd.github.v3.json",
-            "user-agent": "bobril-build-core/1.0.0"
+            "user-agent": "bobril-build-core/1.1.0"
         }
     };
 
@@ -83,7 +83,7 @@ function getDownloadOptions(url: string) {
     var headers = isGitHubUrl
         ? {
             accept: "application/octet-stream",
-            "user-agent": "bobril-build-core/1.0.0"
+            "user-agent": "bobril-build-core/1.1.0"
         }
         : {};
 
@@ -290,9 +290,16 @@ async function checkFreshnessOfCachedLastVersion(): Promise<boolean> {
     let proc = child_process.spawn(
         path.join(toRun, "bb"),
         process.argv.slice(2),
-        { stdio: "inherit" }
+        { stdio: ["pipe", process.stdout, process.stderr] }
     );
     proc.on("exit", (code: number) => {
         process.exit(code);
     });
+    process.stdin.pipe(proc.stdin, { end: true });
+    function endBBCore() {
+        proc.stdin.write("\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b\b" + os.EOL + "quit" + os.EOL);
+    }
+    process.on("SIGINT", endBBCore);
+    process.on("SIGTERM", endBBCore);
+    process.on("SIGBREAK", endBBCore);
 })();